spec/ndt7: clarify optional counter-flow messages
A counter-flow message is a message flowing in the opposite
direction on which binary messages are flowing.

These messages should be used to send measurements regarding
the speed measured at application level. The may possibly
contain additional TCP_INFO measurements (where it's possible).

The previous version of the spec already implied the existence
of such messages. This is just a clarification of wording.

I did some research to understand whether implementing this
feature is doable or not. It turns out that doing this in
measurement-kit/libndt is easy, because we have control over
sockets. So, we can for example, just poll the socket and
see whether we have incoming messages or whether we can send
a message immediately. This allows to embed handling of the
counter-flow messages in the same loop that is already in
use for the current subtest.

I also easily implemented sending counter-flow messages in
m-lab/ndt-server. It seems to me, however, that we're going
to have a more complex implementation of _optionally_
receiving these messages in ndt-server. Currently, I don't
see an easy way of doing that without adding one extra
goroutine, which is a bump of the complexity.